Update: handled some edge cases

+++ automotive/src/main/java/com/akshay/weatherapp/model/WeatherResponseModel.kt
@@ -140,10 +140,10 @@ data class Sys(
     var country: String,
 
     @SerializedName("sunrise")
-    var sunrise: Int,
+    var sunrise: Long,
 
     @SerializedName("sunset")
-    var sunset: Int
+    var sunset: Long
 
 )
